+++ redfishtool/ServiceRoot.py
@@ -3,7 +3,7 @@
 # License: BSD 3-Clause License. For full text see link: https://github.com/DMTF/Redfishtool/LICENSE.md
 
 # redfishtool: ServiceRoot.py
-# v0.9.1
+# v0.9.2
 # contains serviceRoot related subCommands and access functions
 # Class RfServiceRoot
 #  - getServiceRoot   GET /redfish/v1
@@ -36,7 +36,7 @@ class RfServiceRoot:
         if cmdTop is True:  prop=rft.prop
 
         # get root service.  if -P prop, showproperty
-        rc,r,j,d=rft.rftSendRecvRequest(rft.UNAUTHENTICATED_API, 'GET', r.url, relPath=rft.rootPath,prop=prop)
+        rc,r,j,d=rft.rftSendRecvRequest(rft.UNAUTHENTICATED_API, 'GET', rft.rootUri, relPath=rft.rootPath,prop=prop)
 
         #save the rootService response.  The transport may need it later to get link to session and login
         rft.rootResponseDict=d
@@ -59,8 +59,8 @@ class RfServiceRoot:
 
         #calculate relative path = rootPath / odata
         rpath=urljoin(rft.rootPath, "odata")
-
-        rc,r,j,d=rft.rftSendRecvRequest(rft.UNAUTHENTICATED_API, 'GET', r.url, relPath=rpath)
+        
+        rc,r,j,d=rft.rftSendRecvRequest(rft.UNAUTHENTICATED_API, 'GET', rft.rootUri, relPath=rpath)
         if(rc==0 ):
             rft.printVerbose(1," Odata Service Document:",skip1=True, printV12=cmdTop)
         return(rc,r,j,d)
@@ -85,7 +85,7 @@ class RfServiceRoot:
         hdrs = {"Accept": "application/xml", "OData-Version": "4.0" }
 
 
-        rc,r,j,d=rft.rftSendRecvRequest(rft.UNAUTHENTICATED_API, 'GET', r.url, relPath=rpath, jsonData=False, headersInput=hdrs )
+        rc,r,j,d=rft.rftSendRecvRequest(rft.UNAUTHENTICATED_API, 'GET', rft.rootUri, relPath=rpath, jsonData=False, headersInput=hdrs )
         if(rc==0):
             rft.printVerbose(1," Odata Metadata Document:",skip1=True,printV12=cmdTop)
         return(rc,r,j,d)